requirements.txt
openai
chromadb
pdfplumber
tqdm
streamlit
requests
python-dotenv